+++ NEWS.md
@@ -1,9 +1,14 @@
 ClimMobTools 0.3.8 (2021-04-06)
 =========================
 
+### IMPROVEMENTS
+
+* New function `getProjectProgress()` returns the progress in a given project
+
 ### BUG FIXES
 
 * Fix the function to set the URLs for alternative servers in ClimMob
+
 
 ClimMobTools 0.3.7 (2021-03-22)
 =========================